requirements.txt
fastapi==0.115.12
nest-asyncio==1.6.0
openai==1.79.0
python-telegram-bot==22.1
uvicorn==0.34.2
paypalrestsdk==1.14.0
python-jose==3.3.0
python-multipart==0.0.6
pydantic==2.5.2
